Rename animatedgif to animated_gif
