+++ lib/twitter/media/animated_gif.rb
