+++ spec/twitter/media/animated_gif_spec.rb
